LayoutTests/http/tests/security/xssAuditor/form-action-expected.txt

CONSOLE MESSAGE: Refused to execute a JavaScript script. Source code of script found within request.

ALERT: Form action set to about:blank
